+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - Added Craft Requirement tooltip to Recipes
   - If a recipe has a craft requirement, it will now display a small exclamation and tooltip when you hover over it in RRV.
 - Added an option to hide Category Buttons.
+- Fixed not being able to press R or U on pets to get the recipe for them, plus some optimization to recipe lookup.